+++ src/text_input.rs
@@ -43,7 +43,7 @@ impl TextInputState {
 fn draw_text_input(e: &mut DrawEvent) {
     e.ctx.fill_rect(&e.view.bounds, &e.theme.standard.fill);
     e.ctx.stroke_rect(&e.view.bounds, &e.theme.standard.text);
-    let style = TextStyle::new(&e.theme.font, &e.theme.standard.text).with_halign(Align::Start);
+    let style = TextStyle::new(e.theme.font, &e.theme.standard.text).with_halign(Align::Start);
 
     let bounds = e.view.bounds.clone();
     if let Some(state) = e.view.get_state::<TextInputState>() {
@@ -56,8 +56,8 @@ fn draw_text_input(e: &mut DrawEvent) {
                 .stroke_rect(&e.view.bounds.contract(2), &e.theme.standard.text);
             if let Some(state) = e.view.get_state::<TextInputState>() {
                 let n = state.cursor as i32;
-                let w = e.theme.font.character_size.width as i32;
-                let h = e.theme.font.character_size.height as i32;
+                let w = e.theme.font.char_width();
+                let h = e.theme.font.char_height();
                 e.ctx.fill_rect(
                     &Bounds::new(
                         e.view.bounds.position.x + n * w + 5,